Change annotations to contain only provider alias
Motivation
----------
The RFC states that the only annotation on a field should be
an alias for the cryptoprovider and that a provider
should be provided.

Changes
-------
Change how tags are handled and make encryption/decryption
of structs take a map of alias to provider rather than
a key store. Add a register provider function to Transcoder.

Change-Id: I3ef88838f54f66bcc09efe3d1b993db6c9b28094
Reviewed-on: http://review.couchbase.org/96618
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>