Rework how encrypters and decrypters are obtained
Motivation
----------
At present we have providers which fulfill both the encrypter
and decrypter interfaces. This causes a bit of a disconnect
when the user has to create two providers when we they want to
register two encrypters but then only has to register one of them
as a decrypter. Java has implemented this slightly differently
with the provider having EncrypterForKey and Decrypter functions.
We should do this too.

Changes
-------
Rename the legacy providers to decrypters.
Update the aedaes256 provider so that it no longer fulfills the
encrypter or decrypter interfaces.
Add aedaes256 encrypter and decrypter which are accessed via the
provider.

Change-Id: Iaba2c99265751f534131eaa1652eb2c8f3d80488