Use a key provider function for legacy decrypters
Motivation
----------
We currently take a pair of key names in the "constructor" for
the legacy decrypters. This means that the encryption library
is tied to a single encryption key for each legacy decrypter.
Instead of doing this we should take a single function parameter
which returns a private key name for a given public key name. This
will effectively allow us to support multiple encryption keys
for a single decrypter.

Changes
--------
Update the legacy decrypters to take a function parameter instead
of key aliases.

Change-Id: Id2fdd286c857d57f5a952cbee5f3e096d82b30d1